% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/autocrop.R
\name{save_img}
\alias{save_img}
\title{Save image}
\usage{
save_img(cropped, outfile, format, width, units, res, ...)
}
\arguments{
\item{cropped}{}

\item{outfile}{}

\item{format}{}

\item{width}{}

\item{units}{}

\item{res}{}

\item{...}{}
}
\description{
Save image
}
\keyword{internal}

