% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/autocrop.R
\name{autocrop}
\alias{autocrop}
\title{autocrop}
\usage{
autocrop(x, border = 2, outfile = NULL)
}
\arguments{
\item{x}{a file or a raster object}

\item{border}{}

\item{outfile}{}
}
\description{
Automatically crop away white space form an image.  Can read any image
supported by \code{raster}.
}
\examples{
x<-rep(0,49)
x[18]<-1
x[24:25]<-1
x[31:32]<-1
x<-stack(raster(matrix(x,ncol=7,byrow = T)))
extent(x) <- extent(c(1,7,1,7))
x_ac<-autocrop(x,border=0)
x_ac<-autocrop(system.file("extdata/test.tif",package="autocrop"))
}

