% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocrop.R
\name{save_img}
\alias{save_img}
\title{Save image}
\usage{
save_img(cropped, outfile, format, width, units, res, ...)
}
\arguments{
\item{cropped}{x}

\item{outfile}{x}

\item{format}{x}

\item{width}{x}

\item{units}{x}

\item{res}{x}

\item{...}{x}
}
\description{
Save image
}
\keyword{internal}
